# generated/api/PawInHand/rest/model_post.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * PawInHand API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 */

package modules

type Post struct {
	Id string `json:"id,omitempty"`

	Title string `json:"title"`

	Content string `json:"content"`

	AuthorId string `json:"authorId,omitempty"`

	DateCreated string `json:"dateCreated,omitempty"`

	Images []string `json:"images,omitempty"`
}

// AssertPostRequired checks if the required fields are not zero-ed
func AssertPostRequired(obj Post) error {
	elements := map[string]interface{}{
		"title":   obj.Title,
		"content": obj.Content,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertPostConstraints checks if the values respects the defined constraints
func AssertPostConstraints(obj Post) error {
	return nil
}