# resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: PawInHand API
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /ads:
    get:
      description: "Get all animal ads."
      tags:
        - ads
      operationId: GetAllAds
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ad"
        "400":
          description: Malformed or invalid payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                key: "BAD_REQUEST"
                description: "Request payload is either missing or invalid"
        "5XX":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                key: "INTERNAL_SERVER_ERROR"
                description: "An unexpected server error occurred"
    post:
      description: "Create a new ad."
      tags:
        - ads
      operationId: CreateAd
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Ad"
      responses:
        "201":
          description: Ad created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
        "400":
          description: Malformed or invalid payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                key: "BAD_REQUEST"
                description: "Request payload is either missing or invalid"
        "5XX":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                key: "INTERNAL_SERVER_ERROR"
                description: "An unexpected server error occurred"
  /ads/{adId}:
    get:
      description: "Get details of a specific ad by adId."
      tags:
        - ads
      operationId: GetAdById
      parameters:
        - name: adId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
        "400":
          description: Malformed or invalid payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                key: "BAD_REQUEST"
                description: "Request payload is either missing or invalid"
        "410":
          description: City not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                key: "NOT_FOUND"
                description: "The resource you requested cannot be found"
        "5XX":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                key: "INTERNAL_SERVER_ERROR"
                description: "An unexpected server error occurred"
    put:
      description: "Update an ad by adId."
      tags:
        - ads
      operationId: UpdateAdById
      parameters:
        - name: adId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Ad"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
        "400":
          description: Malformed or invalid payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                key: "BAD_REQUEST"
                description: "Request payload is either missing or invalid"
        "410":
          description: City not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                key: "NOT_FOUND"
                description: "The resource you requested cannot be found"
        "5XX":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                key: "INTERNAL_SERVER_ERROR"
                description: "An unexpected server error occurred"
    delete:
      description: "Delete an ad by adId."
      tags:
        - ads
      operationId: DeleteAdById
      parameters:
        - name: adId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Success.
        "400":
          description: Malformed or invalid payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                key: "BAD_REQUEST"
                description: "Request payload is either missing or invalid"
        "5XX":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                key: "INTERNAL_SERVER_ERROR"
                description: "An unexpected server error occurred"
...
  /posts:
    get:
      description: "Get all posts."
      tags:
        - posts
      operationId: GetAllPosts
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "5XX":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                key: "INTERNAL_SERVER_ERROR"
                description: "An unexpected server error occurred"

    post:
      description: "Create a new post."
      tags:
        - posts
      operationId: CreatePost
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Post"
      responses:
        "201":
          description: Post created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Malformed or invalid payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                key: "BAD_REQUEST"
                description: "Request payload is either missing or invalid"
  /posts/{postId}:
    get:
      description: "Get details of a specific post by postId."
      tags:
        - posts
      operationId: GetPostById
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "410":
          description: City not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                key: "NOT_FOUND"
                description: "The resource you requested cannot be found"
    put:
      description: "Update a post by postId."
      tags:
        - posts
      operationId: UpdatePostById
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Post"
      responses:
        "200":
          description: Post updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      description: "Delete a post by postId."
      tags:
        - posts
      operationId: DeletePostById
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Post deleted successfully.
...
components:
  schemas:
...
    Ad:
      type: object
      required:
        - title
        - description
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        shelterId:
          type: string
        dateCreated:
          type: string
          format: date
        images:
          type: array
          items:
            type: string
...
    Post:
      type: object
      required:
        - title
        - content
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        authorId:
          type: string
        dateCreated:
          type: string
          format: date
        images:
          type: array
          items:
            type: string
...
    Error:
      type: object
      required:
        - key
        - description
      properties:
        key:
          type: string
          example: "BAD_REQUEST"
        description:
          type: string
          example: "Request payload is either missing or invalid"